requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
# PostgreSQL (optional - for production)
psycopg2-binary==2.9.9; extra == "postgres"
# SQLite (default - for local development)
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# File Processing
pdfplumber==0.10.3
python-docx==1.1.0
docx2txt==0.8

# NLP & Text Processing
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2
rank-bm25==0.2.2
rapidfuzz==3.5.2

# AI & LLM
langchain==0.1.0
langchain-openai==0.0.2
langchain-google-genai==0.0.6
langchain-community==0.0.10
langgraph==0.0.20
openai==1.6.1
google-generativeai==0.3.2

# Vector Store & Embeddings
faiss-cpu==1.7.4
chromadb==0.4.18
sentence-transformers==2.2.2

# Frontend
streamlit==1.29.0
plotly==5.17.0
pandas==2.1.4

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
pydantic-extra-types==2.2.0

# Development & Debugging
langsmith==0.0.69
pytest==7.4.3
pytest-asyncio==0.21.1
